+++ library-common/src/main/java/universum/studios/android/fragment/annotation/ActionModeOptions.java
@@ -20,6 +20,7 @@ package universum.studios.android.fragment.annotation;
 
 import android.support.annotation.MenuRes;
 
+import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -30,6 +31,7 @@ import java.lang.annotation.Target;
  *
  * @author Martin Albedinsky
  */
+@Documented
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface ActionModeOptions {