+++ library-manage-base/src/main/java/universum/studios/android/fragment/annotation/FactoryFragments.java
@@ -18,6 +18,7 @@
  */
 package universum.studios.android.fragment.annotation;
 
+import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -31,6 +32,7 @@ import universum.studios.android.fragment.manage.FragmentFactory;
  *
  * @author Martin Albedinsky
  */
+@Documented
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface FactoryFragments {