+++ library-manage-base/src/main/java/universum/studios/android/fragment/manage/BaseFragmentFactory.java
@@ -24,7 +24,6 @@ import android.support.annotation.Nullable;
 import android.text.TextUtils;
 import android.util.SparseArray;
 
-import universum.studios.android.fragment.FragmentsConfig;
 import universum.studios.android.fragment.annotation.FactoryFragment;
 import universum.studios.android.fragment.annotation.FactoryFragments;
 import universum.studios.android.fragment.annotation.FragmentAnnotations;
@@ -64,7 +63,7 @@ import universum.studios.android.fragment.annotation.handler.FragmentFactoryAnno
  */
 public abstract class BaseFragmentFactory implements FragmentFactory {
 
-	/**
+	/*
 	 * Constants ===================================================================================
 	 */
 
@@ -73,15 +72,15 @@ public abstract class BaseFragmentFactory implements FragmentFactory {
 	 */
 	// private static final String TAG = "BaseFragmentFactory";
 
-	/**
+	/*
 	 * Interface ===================================================================================
 	 */
 
-	/**
+	/*
 	 * Static members ==============================================================================
 	 */
 
-	/**
+	/*
 	 * Members =====================================================================================
 	 */
 
@@ -108,22 +107,22 @@ public abstract class BaseFragmentFactory implements FragmentFactory {
 	 */
 	private boolean mFragmentProvided;
 
-	/**
+	/*
 	 * Constructors ================================================================================
 	 */
 
 	/**
 	 * Creates a new instance of BaseFragmentFactory.
 	 * <p>
-	 * If annotations processing is enabled via {@link FragmentsConfig} all annotations supported by
-	 * this class will be processed/obtained here so they can be later used.
+	 * If annotations processing is enabled via {@link FragmentAnnotations} all annotations supported
+	 * by this class will be processed/obtained here so they can be later used.
 	 */
 	public BaseFragmentFactory() {
 		this.mAnnotationHandler = onCreateAnnotationHandler();
 		this.mItems = mAnnotationHandler == null ? null : mAnnotationHandler.getFragmentItems();
 	}
 
-	/**
+	/*
 	 * Methods =====================================================================================
 	 */
 
@@ -242,7 +241,7 @@ public abstract class BaseFragmentFactory implements FragmentFactory {
 		return mItems.indexOfKey(fragmentId) >= 0 ? mItems.get(fragmentId).tag : createFragmentTag(getClass(), Integer.toString(fragmentId));
 	}
 
-	/**
+	/*
 	 * Inner classes ===============================================================================
 	 */
 }