+++ samples/src/main/java/universum/studios/android/samples/fragment/ui/welcome/WelcomeActivity.java
@@ -59,10 +59,10 @@ public final class WelcomeActivity extends SamplesActivity implements FragmentRe
 	public Fragment interceptFragmentRequest(@NonNull FragmentRequest request) {
 		switch (request.fragmentId()) {
 			case WelcomeFragments.SIGN_IN:
-				request.transition(FragmentTransitions.CROSS_FADE_AND_HOLD).addToBackStack(true);
+				request.transition(FragmentTransitions.SLIDE_TO_LEFT).addToBackStack(true);
 				break;
 			case WelcomeFragments.SIGN_UP:
-				request.transition(FragmentTransitions.CROSS_FADE).addToBackStack(true);
+				request.transition(FragmentTransitions.SLIDE_TO_RIGHT).addToBackStack(true);
 				break;
 			case WelcomeFragments.LOST_PASSWORD:
 				request.transition(FragmentTransitions.CROSS_FADE).addToBackStack(true);